+++ lib/utils/option_async_value.dart
@@ -1,0 +1,11 @@
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:oxidized/oxidized.dart';
+
+extension OptionAsyncValueFlattener<T> on Option<AsyncValue<T>> {
+  AsyncValue<T> flat() {
+    return when(
+      some: (value) => value,
+      none: () => const AsyncLoading(),
+    );
+  }
+}
